my_app_v3/requirements.txt
Flask>=1.1.2
google-cloud-storage>=1.18.1
requests>=2.23.0